moved a few more things out to main.py
